Update: cypress command
- Removing LR page object and instead adding it as a command
